Fix show_clue to deal with duplicates a different way
Do a pass that checks for 'correct colour and position',
then do a pass that checks for 'correct colour in the wrong position'.

Previous algorithm let 'correct colour in the wrong position'
steal a count value, so if there was a 'correct colour and position'
to the right of the 'correct colour in the wrong position',
it would never get a clue ('x'), but there would be an ('o') clue
to the left of this.