Improve computer guessing
MindReadAlgorithm has an extra restriction on what new number
it assigns to code pegs with an empty clue, if that code peg
has been 'pushed' by a partial peg that has been 'moved'.

The algorithm will check that the 'pushed' code peg does not get
assigned the same number as the last code peg in that position,
if that code peg had been given a 'partial' clue in the previous guess.

This is because a 'partial' clue is saying that, that peg, should
be 'moved', so that number should not exist at that position.